[front] enh: rework the Pod Apps tab into a tile grid
